bosminer: api: Move sources to separate crate
- create new crate 'ii-cgminer-api' in 'open/protocols'
