+++ open/bosminer/bosminer/Cargo.toml
@@ -13,15 +13,12 @@ ii-logging = { path = "../../utils-rs/logging" }
 ii-stats = { path = "../../utils-rs/stats" }
 ii-stratum = { path = "../../protocols/stratum" }
 ii-wire = { path = "../../protocols/wire" }
-async-trait = "0.1.13"
+async-trait = "0.1"
 failure = "0.1.5"
 once_cell = "1.2"
 downcast-rs = "1.0.4"
 clap = "2.33.0"
 hex = "0.3.1"
 git-version = "0.3.3"
-serde = { version = "1.0", features = ["derive"] }
-serde_derive = "1.0"
-serde_repr = "0.1"
-serde_json = "1.0"
+serde = "1.0"
 config = "0.9"
