+++ open/protocols/cgminer-api/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "ii-cgminer-api"
+version = "0.1.0"
+authors = ["Braiins <braiins@braiins.com>"]
+license = "GPL-3.0-or-later"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ii-async-compat = { path = "../../utils-rs/async-compat" }
+ii-logging = { path = "../../utils-rs/logging" }
+ii-wire = { path = "../../protocols/wire" }
+async-trait = "0.1"
+serde = { version = "1.0", features = ["derive"] }
+serde_repr = "0.1"
+serde_json = "1.0"
