+++ open/protocols/cgminer-api/src/command.rs
