+++ open/protocols/cgminer-api/src/response.rs
