+++ open/protocols/cgminer-api/src/server.rs
