+++ open/protocols/cgminer-api/src/support.rs
