+++ open/protocols/cgminer-api/src/test.rs
